Moved dice to new folder. Renamed constant to be clear.
