+++ backend/src/game/constants.py
@@ -1,2 +1,2 @@
 DEFAULT_HEALTH = 10
-ZERO = 0
+DEATH_HIT_POINT = 0
