+++ backend/src/game/player.py
@@ -7,7 +7,7 @@ class Player:
         self.maximum_health = self.current_health = constants.DEFAULT_HEALTH
         self.location = Locations.OUTSIDE
         self.is_alive = True
-        self.victory_points = constants.ZERO
+        self.victory_points = constants.DEATH_HIT_POINT
 
     def move_to_tokyo(self):
         self.location = Locations.TOKYO
@@ -19,10 +19,10 @@ class Player:
         self.current_health += change_integer
         if self.current_health > self.maximum_health:
             self.current_health = self.maximum_health
-        if self.current_health <= constants.ZERO:
+        if self.current_health <= constants.DEATH_HIT_POINT:
             self.is_alive = False
 
     def change_victory_points(self, change_integer):
         self.victory_points += change_integer
-        if self.victory_points < constants.ZERO:
-            self.victory_points = constants.ZERO
+        if self.victory_points < constants.DEATH_HIT_POINT:
+            self.victory_points = constants.DEATH_HIT_POINT
